ref: add id variable to message generators, move get_code_id to its own file in common
